Popup Options and Last Check In
Users can now select what go to buttons appear in the popup through the
options page.

A system to save the last checked in item has also been added that is
now also a go to button.

Popup also updated to look more appealing. Before release will be
changed again however to be more inkeeping with the Trakt house style.